feat: move error handlers to file
