+++ .gitignore
@@ -2,4 +2,4 @@ node_modules
 .idea
 out
 build
-*.log
+log
